Improve session/user workflow
* Include the user information in the AuthWrapper
* Have the router provide the UserContext
* Use only /api/user for session validation
